+++ backend/src/main/java/net/primepro/primepro/repository/UsersRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface UsersRepo extends JpaRepository<OurUsers, Integer> {
+public interface UsersRepo extends JpaRepository<OurUsers, Long> {
 
     Optional<OurUsers> findByEmail(String email);
 }
